bummer tests need these, tweak later

+++ t.sh
@@ -108,11 +108,16 @@ esac
 #neato -Tpng a.gv >n.png
 ##[[ -s n.png ]] && [[ $silent = 0 ]] && open n.png
 
-rm $OUT
-#Compress
-which  pngquant >/dev/null
-if [ $? -eq 0 ]; then
-  pngquant --ext .png --force --speed 1 --quality 0-10 $png
+#When running tests, these are important
+test=1
+if [ $test -eq 1 ]; then
+  rm $OUT
+  #Compress
+  which  pngquant >/dev/null
+  if [ $? -eq 0 ]; then
+    pngquant --ext .png --force --speed 1 --quality 0-10 $png
+  fi
+  echo $png
 fi
-echo $png
+
 exit 0
